Extract Divider component from CommonStyles file
