+++ src/stories/Divider.stories.js
@@ -1,5 +1,5 @@
 import React from "react";
-import { Divider } from "components/CommonStyles";
+import Divider from "components/Divider";
 
 export default {
   component: Divider,
